requirements.txt
streamlit
google-generativeai
python-dotenv
watchdog
edge-tts
